renaming files to have simple naming convention. renaming the import function to make it more simple. chaning all links and connections
