+++ libraries/base.rb
@@ -1,7 +1,7 @@
 #
 # Author:: Sander Botman <sbotman@schubergphilis.com>
 # Cookbook Name:: nagios
-# Library:: nagios_base
+# Library:: base
 #
 # Copyright 2014, Sander Botman
 #
