+++ libraries/command.rb
@@ -1,7 +1,7 @@
 #
 # Author:: Sander Botman <sbotman@schubergphilis.com>
 # Cookbook Name:: nagios
-# Library:: nagios_command
+# Library:: command
 #
 # Copyright 2014, Sander Botman
 #
@@ -17,7 +17,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-require_relative 'nagios_base'
+require_relative 'base'
 
 class Nagios
   class Command < Nagios::Base
@@ -41,7 +41,7 @@ class Nagios
       self.command_name
     end
 
-    def import_hash(hash)
+    def import(hash)
       update_options(hash)
     end
