+++ libraries/servicedependency.rb
@@ -1,7 +1,7 @@
 #
 # Author:: Sander Botman <sbotman@schubergphilis.com>
 # Cookbook Name:: nagios
-# Library:: nagios_servicedependency
+# Library:: servicedependency
 #
 # Copyright 2014, Sander Botman
 #
@@ -17,7 +17,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-require_relative 'nagios_base'
+require_relative 'base'
 
 class Nagios
   class Servicedependency < Nagios::Base
@@ -79,7 +79,7 @@ class Nagios
       self.service_description
     end
 
-    def import_hash(hash)
+    def import(hash)
       update_options(hash)
       update_members(hash, 'host_name', Nagios::Host)
       update_members(hash, 'hostgroup_name', Nagios::Hostgroup)
